+++ src/BeadMap.cpp
@@ -6,25 +6,25 @@
 #include "BeadMap.hpp"
 
 // Given a bitmap and a position in that, clear this slot (set to 0)
-inline uint32_t clear_slot(uint32_t bitmap, uint8_t pos) {
+inline uint16_t clear_slot(uint16_t bitmap, uint8_t pos) {
     return bitmap & ~(1 << pos);
 }
 
 // Given a bitmap and a position in that, set this slot (set to 1)
 // This will indicate that this position in the bead array contains a bead
-inline uint32_t set_slot(uint32_t bitmap, uint8_t pos) {
+inline uint16_t set_slot(uint16_t bitmap, uint8_t pos) {
     return bitmap | (1 << pos);
 }
 
 // Given a bitmap and a position in that, report whether there is a bead in
 // this slot
-inline bool is_slot_set(uint32_t bitmap, uint8_t pos) {
+inline bool is_slot_set(uint16_t bitmap, uint8_t pos) {
     return bitmap & (1 << pos);
 }
 
 // Given a bitmap, return the position of the next occupied slot
-inline uint8_t get_next_slot(uint32_t bitmap) {
-    uint32_t mask = 0x1;
+inline uint8_t get_next_slot(uint16_t bitmap) {
+    uint16_t mask = 0x1;
     for(int i = 0; i < MAX_BEADS; i++) {
         if(bitmap & mask) {
             return i;
@@ -35,8 +35,8 @@ inline uint8_t get_next_slot(uint32_t bitmap) {
 }
 
 // Given a bitmap, return the position of the next empty slot
-inline uint8_t get_next_free_slot(uint32_t bitmap) {
-    uint32_t mask = 0x1;
+inline uint8_t get_next_free_slot(uint16_t bitmap) {
+    uint16_t mask = 0x1;
     for(int i = 0; i < MAX_BEADS; i++){
         if(!(bitmap & mask)) {
            return i;
@@ -47,9 +47,9 @@ inline uint8_t get_next_free_slot(uint32_t bitmap) {
 }
 
 // Given a bitmap, print the occupancy to the screen
-inline void print_bitmap(uint32_t bitmap) {
+inline void print_bitmap(uint16_t bitmap) {
     printf("bitmap = ");
-    uint32_t mask = 0x1;
+    uint16_t mask = 0x1;
     for(int i = 0; i < MAX_BEADS; i++) {
         if(bitmap & mask) {
             printf("1");
@@ -63,9 +63,9 @@ inline void print_bitmap(uint32_t bitmap) {
 
 // Given a bitmap, count the number of occupied slots and therefore, the number
 // of beads this bitmap represents
-inline uint8_t get_num_beads(uint32_t bitmap) {
+inline uint8_t get_num_beads(uint16_t bitmap) {
     uint8_t cnt = 0;
-    uint32_t mask = 0x1;
+    uint16_t mask = 0x1;
     for(int i = 0; i < MAX_BEADS; i++) {
         if(bitmap & mask) {
             cnt++;
